ai_search/backend/requirements.txt
# AI Search Engine Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML Libraries
openai==1.3.7
anthropic==0.3.11
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0

# Vector Database & Search
pgvector==0.2.4
psycopg2-binary==2.9.7
faiss-cpu==1.7.4
elasticsearch==8.11.0
redis==5.0.1

# Text Processing
beautifulsoup4==4.12.2
html2text==2020.1.16
spacy==3.7.2
nltk==3.8.1
langchain==0.0.330

# Data & Utils
pandas==2.1.3
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pydantic-settings==2.0.3

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
